Respond to feedback on aarch64 interrupt code
