+++ aarch64/src/trap.S
@@ -25,8 +25,8 @@
 	b	\label
 .endm
 
-// Aarch64 calling convention dictates that we save:
-// - x0-18
+// We save all general purpose registers, and a few useful system registers:
+// - x0-28
 // - x29 (Frame pointer)
 // - x30 (Link register)
 // We additionally save the registers below, so that they're avaialble via the InterruptFrame
@@ -34,7 +34,7 @@
 // - ELR_EL1 (Exception link register EL1)
 // - FAR_EL1 (Fault address register EL1)
 .macro handle_interrupt type
-	sub 	sp, sp, #208
+	sub 	sp, sp, #288
 
 	// Caller-saved registers, FP
 	stp 	x0, x1, [sp, #16 * 0]
@@ -46,24 +46,29 @@
 	stp	x12, x13, [sp, #16 * 6]
 	stp	x14, x15, [sp, #16 * 7]
 	stp	x16, x17, [sp, #16 * 8]
-	stp	x18, x29, [sp, #16 * 9]
+	stp	x18, x19, [sp, #16 * 9]
+	stp	x20, x21, [sp, #16 * 10]
+	stp	x22, x23, [sp, #16 * 11]
+	stp	x24, x25, [sp, #16 * 12]
+	stp	x26, x27, [sp, #16 * 13]
+	stp	x28, x29, [sp, #16 * 14]
 
 	// LR, ESR_EL1
 	mrs	x0, esr_el1
-	stp	x30, x0, [sp, #16 * 10]
+	stp	x30, x0, [sp, #16 * 15]
 
 	// ELR_EL1, FAR_EL1
 	mrs	x1, elr_el1
 	mrs	x2, far_el1
-	stp	x1, x2, [sp, #16 * 11]
+	stp	x1, x2, [sp, #16 * 16]
 
-	// Interrupt type, padding
+	// Interrupt type
 	ldr	x3, =\type
-	stp	x3, xzr, [sp, #16 * 12]
+	str	x3, [sp, #16 * 17]
 
 	// Pass pointer to InterruptFrame (on stack) as the first arg
 	mov	x0, sp
-	bl	handle_exception_with_ptr
+	bl	trap_unsafe
 
 	// Restore caller-saved registers
 	ldp	x0, x1, [sp, #16 * 0]
@@ -75,11 +80,16 @@
 	ldp	x12, x13, [sp, #16 * 6]
 	ldp	x14, x15, [sp, #16 * 7]
 	ldp	x16, x17, [sp, #16 * 8]
-	ldp	x18, x29, [sp, #16 * 9]
-	ldp	x30, xzr, [sp, #16 * 10]
+	ldp	x18, x19, [sp, #16 * 9]
+	ldp	x20, x21, [sp, #16 * 10]
+	ldp	x22, x23, [sp, #16 * 11]
+	ldp	x24, x25, [sp, #16 * 12]
+	ldp	x26, x27, [sp, #16 * 13]
+	ldp	x28, x29, [sp, #16 * 14]
+	ldr	x30, [sp, #16 * 14]
 
-	// Remaining stack frame is registers we ignore
-	add	sp, sp, #208
+	// Remaining stack frame consists of systems registers we can just ignore
+	add	sp, sp, #288
 
 	eret
 .endm
